backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
twilio
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
streamlit